requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
jmespath==1.0.1
nextcloud-api-wrapper==0.2.3
requests==2.28.1
six==1.16.0
urllib3==1.26.12
